Move editor size calculation into lapce-ui
